+++ BasicPythonExercisesAndTrainings/01a-Excercices.py
@@ -3029,13 +3029,9 @@ print(f'Quantity of Grades for all students without pets {QuantityOfGradesForAll
 print(f'Average of Grades for all students without pets {AverageGradesForAllStudentsWithoutPets}')
 
 
+# Write a loop that adds one to every number in the list of numbers
+
 # In[194]:
-
-
-#Write a loop that adds one to every number in the list of numbers
-
-
-# In[195]:
 
 
 numbers = [5, 6, 7, 8, 9, 10]
@@ -3043,7 +3039,7 @@ number = numbers[0] + 1
 number
 
 
-# In[196]:
+# In[195]:
 
 
 AddedNumbers = []
@@ -3054,7 +3050,7 @@ for number in numbers:
 AddedNumbers
 
 
-# In[197]:
+# In[196]:
 
 
 AddedNumbers = [number + 1 for number in numbers]
@@ -3064,821 +3060,3 @@ AddedNumbers
 # Created by [Ryan Orsinger](https://ryanorsinger.com)
 # 
 # Source code on [https://github.com/ryanorsinger/101-exercises](https://github.com/ryanorsinger/101-exercises)
-
-# # For Loop
-
-# #### For Loop - It executes a block of code repeatedly
-
-# In[198]:
-
-
-for number in range(5):
-    print("Thank you")
-
-
-# In[199]:
-
-
-for x in [0,1,2]:
-    print(x)
-
-
-# In[260]:
-
-
-symbols = {'12123312123': 'APPL',
-           '765757567756':'CSVO',
-           '598918919':'ORCL'}
-
-for k in symbols:
-    print(symbols[k])
-
-
-# In[261]:
-
-
-for key, value in symbols.items():
-    print(key, value)
-
-
-# In[201]:
-
-
-for x in 'ORACLE':
-    print(x)
-
-
-# In[202]:
-
-
-languages = ['R', 'Python', 'Scala', 'Java',' Julia']
-
-
-# In[203]:
-
-
-for index in range(len(languages)):
-    print("Current language:", languages[index])
-
-
-# In[204]:
-
-
-for index in range(5):
-    print("Current language", languages[index])
-
-
-# In[205]:
-
-
-#Printing Thank you a 3 times
-
-for number in range(3):
-    print("Thank you")
-
-
-# In[263]:
-
-
-# it stops the iteration if we encounter 5
-for i in range(10):
-    print(i)
-    if i == 5:
-        break
-
-
-# In[264]:
-
-
-for i in range(20):
-    if i % 2 == 0:
-        continue
-    print(i)
-
-
-# In[206]:
-
-
-# we ignore value for example 2
-for x in [0,1,2,3]:
-    if x == 2:
-        continue
-    print(x)
-
-
-# In[3]:
-
-
-nums = [1, 2, 3, 4, 5]
-
-for num in nums:
-    if num == 3:
-        print('Found')
-        break
-    print(num)
-
-
-# In[4]:
-
-
-nums = [1, 2, 3, 4, 5]
-
-for num in nums:
-    if num == 3:
-        print('Skip 3')
-        continue
-    print(num)
-
-
-# #### Nested loops
-
-# In[207]:
-
-
-for number in range(3):
-    print("---------------------------------------------")
-    print("I am outer loop iteration "+str(number))
-    #inner loop
-    for another_number in range(5):
-        print("*************************")
-        print("I am inner loop iteration "+str(another_number))
-
-
-# In[208]:
-
-
-for i in range(5):
-    print('*')
-    for j in range(i + 1):
-        print("*",end=' ')
-
-for i in range(5,-1,-1):
-    print('*')
-    for j in range(i + 1):
-        print("*",end=' ')
-       
-
-
-# In[209]:
-
-
-for i in range(2):
-    print('outer',i)
-
-
-# In[210]:
-
-
-for i in range(4):
-    print('outer',i)
-    for j in range(i + 1):
-        print('inner',j)
-
-
-# In[211]:
-
-
-for i in range(4):
-    print('outer',i)
-    for j in range(i + 1):
-        print('inner',j, end=" ")
-
-
-# In[212]:
-
-
-for i in range(4,-1,-1):
-    print('outer',i)
-    for j in range(i + 1):
-        print('inner',j, end=" ")
-
-
-# In[213]:
-
-
-for i in range(5):
-    print("*")
-    for j in range(i+1):
-        print("*", end=" ")
-
-
-# In[214]:
-
-
-for i in range(5,-1,-1):
-    print("*")
-    for j in range(i + 1):
-        print("*", end=" ")
-
-
-# In[265]:
-
-
-for i in range(3):
-    for j in range(3):
-        print(i, j )
-
-
-# In[267]:
-
-
-for index, value in enumerate(['a', 'b', 'c']):
-    print(index, value)
-
-
-# In[6]:
-
-
-nums = [1, 2, 3, 4, 5]
-
-for num in nums:
-    for letter in 'abc':
-        print(num, letter) #for each number it will loop through every character in this string 'abc' 
-                            #and print out the number and character, then move on to the next number and do it over again
-
-
-# ### Loop
-
-# In[215]:
-
-
-for i in range(10):
-    print(i+1)
-
-
-# In[216]:
-
-
-for i in 'Datacamp':
-    if i == 'a':
-        print(i)
-
-
-# In[217]:
-
-
-sequence = [1,2,8,100,200,'datacamp','tutorial']
-
-
-# In[218]:
-
-
-for i in sequence:
-    print(i)
-
-
-# In[219]:
-
-
-len(sequence)
-
-
-# In[220]:
-
-
-sequence[5]
-
-
-# In[221]:
-
-
-for i in range(len(sequence)):
-    print(sequence[i])
-
-
-# In[222]:
-
-
-for i in range(len(sequence)):
-    element = sequence[i]
-    if type(element) == int:
-        sequence[i] = element  + 1
-
-
-# In[223]:
-
-
-sequence
-
-
-# In[224]:
-
-
-for i in range(1,20,2):
-    print(i)
-
-
-# In[225]:
-
-
-for i in range(11):
-    for j in range(i):
-        print(i, end= ' ')
-    print()
-
-
-# In[226]:
-
-
-import pandas as pd
-iris = pd.read_csv('https://raw.githubusercontent.com/mwaskom/seaborn-data/master/iris.csv')
-
-
-# In[227]:
-
-
-iris.head()
-
-
-# In[228]:
-
-
-len(iris) # rows
-
-
-# In[229]:
-
-
-iris.iloc[1,2]
-
-
-# In[230]:
-
-
-iris.iloc[111,4] # we are on the 'species' column
-
-
-# In[231]:
-
-
-for i in range(len(iris)):
-    Class = iris.iloc[i,4]
-    if Class == 'versicolor' and i < 70:
-        print(Class)
-
-
-# In[232]:
-
-
-for index, row in iris.iterrows():
-    print(row["species"])
-
-
-# In[233]:
-
-
-iris.at[0,'sepal_length']
-
-
-# In[234]:
-
-
-columns =  ['sepal_length','sepal_width','petal_length','petal_width']
-
-for indices, row in iris.iterrows():
-    for column in columns:
-        iris.at[indices, column] = row[column] + 2
-
-
-# In[235]:
-
-
-iris
-
-
-# ### Lambda function to iterate
-
-# In[236]:
-
-
-x = [20, 30, 40, 50, 60]
-y = []
-
-
-# In[237]:
-
-
-x
-
-
-# In[238]:
-
-
-for v in x:
-    y += [v * 5]
-
-
-# In[239]:
-
-
-x
-
-
-# In[240]:
-
-
-y = map(lambda x: x * 5, x)
-y
-
-
-# In[241]:
-
-
-list(y)
-
-
-# In[242]:
-
-
-columns
-
-
-# In[243]:
-
-
-for column in columns:
-    iris[column] = iris[column].apply(lambda row: row + 2)
-
-
-# In[244]:
-
-
-iris
-
-
-# # While
-
-# #### While loop - it executes a piece of code over and over again while a given condition still holds true
-
-# In[245]:
-
-
-number = 2
-
-while number < 5:
-    print('Thank you')
-    number += 1 
-
-
-# In[246]:
-
-
-number = 0
-while number < 5:
-    print(f"The number is {number}")
-    number += 1 
-
-
-# In[247]:
-
-
-number = 0
-
-while number < 10:
-    if number % 2 == 0:
-        print("The number "+str(number)+" is even")
-    else:
-        print("The number "+str(number)+" is odd")
-        
-    number += 1
-    
-
-
-# In[248]:
-
-
-number = 2
-
-while number < 6:
-    print(number)
-    number += 1
-
-
-# In[249]:
-
-
-def collatz(number):
-    if number % 2 == 0:
-        print(number // 2)
-        return number // 2
-    elif number % 2 == 1:
-        result = 3 * number + 1 
-        print(result)
-        return result 
-    
-n = input("Give me a number: ")
-
-while n !=1:
-    n = collatz(int(n))
-
-
-# In[250]:
-
-
-number = 0
-
-while number < 3:
-    print("Thank you ")
-    number += 1
-
-
-# In[251]:
-
-
-number = 1
-
-while number < 10:
-    while number % 2 ==0:
-        print("The number "+str(number)+" is even")
-        break 
-    number += 1
-
-
-# In[252]:
-
-
-for number in range(3) : 
-    print("-------------------------------------------")
-    print("I am outer loop iteration "+str(number))
-    for another_number in range(3):
-        print("****************************")
-        print("I am inner loop iteration "+str(another_number))
-        break
-
-
-# In[253]:
-
-
-error = 50.0
-
-while error > 1:
-    error = error / 4
-    print(error)
-
-
-# In[254]:
-
-
-offset = 8
-
-while offset != 0:
-    print("Correcting")
-    offset -= 1
-    print(offset)
-    
-
-
-# In[255]:
-
-
-while True:
-    user_input = input("Enter a positive number: ")
-    if user_input.isdigit() and int(user_input) > 0:
-        break
-    print("Invalid input. Please try again")
-
-
-# In[256]:
-
-
-stock = 10 
-num_of_purchases = 0
-
-while num_of_purchases < stock:
-    num_of_purchases += 1
-    if stock - num_of_purchases > 7:
-        print("Plenty of stock remaining")
-    elif stock - num_of_purchases > 3:
-        print("Some stock remaining")
-    elif stock - num_of_purchases != 0:
-        print("Low stock!")
-    else:
-        print("No stock")
-        
-    
-
-
-# In[9]:
-
-
-x = 0 
-
-while x < 10:
-    print(x)
-    x += 1
-
-
-# In[11]:
-
-
-x = 0
-
-while x < 10:
-    if x == 5:
-        break
-    print(x)
-    x += 1
-
-
-# In[12]:
-
-
-x = 0 
-
-while True:
-    if x == 5:
-        break
-    x += 1
-    print(x)
-
-
-# # range()
-
-# In[257]:
-
-
-range(10)
-
-
-# In[258]:
-
-
-for i in range(10):
-    print(i)
-
-
-# In[259]:
-
-
-for i in range(5,10,2):
-    print(i)
-
-
-# In[8]:
-
-
-for i in range(1,11):
-    print(i)
-
-
-# In[13]:
-
-
-type(range(100))
-
-
-# In[15]:
-
-
-list(range(10))
-
-
-# In[16]:
-
-
-len(list(range(10)))
-
-
-# In[17]:
-
-
-for seq in range(-5,0):
-    print(seq)
-
-
-# In[18]:
-
-
-for seq in range(50,1000,100):
-    print(seq)
-
-
-# In[19]:
-
-
-for seq in range(100,10,-10):
-    print(seq)
-
-
-# In[69]:
-
-
-list1 = [2,4,6,8,10,12,14,16,18,20]
-counter = 0
-
-for i in range(len(list1)):
-    counter = counter + list1[i]
-    print(counter)
-print(f'Sum of the list : {counter}')
-
-SumOfTheList = sum([x for x in list1])
-print(f'List Comprehension SumOfTheList: {SumOfTheList}')
-
-
-# In[42]:
-
-
-from itertools import chain 
-
-a1 = range(10,0, -2)
-a2 = range(30,20,-2)
-a3 = range(50,40,-2)
-final = chain(a1, a2, a3)
-print(final)
-print(list(final))
-
-
-# In[48]:
-
-
-for a in final:
-    print(a)
-
-
-# In[54]:
-
-
-a = list(range(0, 10, 3))
-b = list(range(0, 11, 3))
-c = list(range(0, 11, 2))
-print(a == b)
-print(a == c)
-
-
-# # if elif else
-
-# In[70]:
-
-
-z = 4 
-
-if z % 2 == 0:
-    print('z is even')
-
-
-# In[72]:
-
-
-if z % 2 == 0:
-    print('Checking ' + str(z))
-    print('z is even')
-
-
-# In[73]:
-
-
-z = 5 
-if z % 2 == 0: # False, no output
-    print('Checking ' + str(z))
-    print('z is even')
-    
-
-
-# In[77]:
-
-
-z = 5 
-
-if z % 2 == 0:
-    print('z is even')
-else:
-    print('z is odd')
-
-
-# In[87]:
-
-
-z = 5
-
-if z % 2 == 0:
-    print('z is divisible by 2')
-elif z % 3 == 0:
-    print('z is divisible by 3')
-else:
-    print('z is neither divisible by 2 nor by 3')
-
-
-# In[88]:
-
-
-room = 'bed'
-area = 14.0
-
-if room == 'kit':
-    print('Looking around in the kitchen')
-elif room == 'bed':
-    print('Looking around in the bedroom')
-else:
-    print('Looking around elsewhere')
-    
-if area > 15:
-    print('It is a small place')
-else:
-    print('Pretty small')
-
-
-# In[ ]:
-
-
-
-
